// v1/web/generated/api/model/recurringTransaction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cashmemo V1 API
 * OpenAPI spec version: 1.0.0
 */
import type { Cadence } from './cadence';
import type { CurrencyCode } from './currencyCode';
import type { RecurringStatus } from './recurringStatus';
import type { TransactionDirection } from './transactionDirection';

export interface RecurringTransaction {
  amount: string;
  category_id: string;
  currency: CurrencyCode;
  direction: TransactionDirection;
  frequency: Cadence;
  id: string;
  next_due_date: string;
  /** @nullable */
  note?: string | null;
  start_date: string;
  status: RecurringStatus;
  wallet_id: string;
}